move loadSearchRecordsMock to mock-actions.ts file
